+++ baekjoon/.gitignore
@@ -52,3 +52,4 @@
 /boj_5052$Node.class
 /boj_2011.class
 /boj_16510.class
+/boj_3085.class
